Refactor filter
