+++ backend/apps/iamstudent/filters/student_filter.py
@@ -1,10 +1,18 @@
-import django.forms as forms
+from crispy_forms.bootstrap import InlineRadios
+from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Column, Div, HTML, Layout, Row
+from django import forms
 from django.utils.translation import gettext_lazy as _
 import django_filters as filters
 
-from apps.iamstudent.forms import form_labels, get_form_helper_filter
+from apps.iamstudent.forms.student import button_group_filter, form_labels
 from apps.iamstudent.models import Student
-from apps.iamstudent.models.student import AUSBILDUNGS_TYPEN_COLUMNS, BEZAHLUNG_CHOICES_Filter
+from apps.iamstudent.models.student import (
+    AUSBILDUNGS_IDS,
+    AUSBILDUNGS_TYPEN,
+    AUSBILDUNGS_TYPEN_COLUMNS,
+    BEZAHLUNG_CHOICES_Filter,
+)
 
 
 class StudentJobRequirementsFilter(filters.FilterSet):
@@ -177,3 +185,135 @@ class StudentJobRequirementsFilter(filters.FilterSet):
                     self.form.fields[a_field].label = form_labels[a_field]
 
             self.form_helper = get_form_helper_filter()
+
+
+def get_form_helper_filter():
+    helper = FormHelper()
+    helper.form_id = "id-exampleForm"
+    helper.form_class = "blueForms"
+    helper.form_method = "get"
+
+    helper.form_action = "submit_survey"
+    helper.form_style = "inline"
+    helper.layout = Layout(
+        Div(
+            Row(
+                *[
+                    Column(
+                        "ausbildung_typ_%s" % k.lower(),
+                        css_class="ausbildung-checkbox form-group col-md-6 mb-0",
+                        css_id="ausbildung-checkbox-%s" % AUSBILDUNGS_IDS[k],
+                    )
+                    for k in AUSBILDUNGS_TYPEN.keys()
+                ]
+            ),
+            css_id="div-berufsausbildung-dropdown",
+        )
+    )
+
+    for ausbildungstyp, felder in AUSBILDUNGS_TYPEN.items():
+        if len(felder) != 0:
+            if ausbildungstyp != "MEDSTUD":
+                helper.layout.extend(
+                    [
+                        Div(
+                            HTML(
+                                "<hr><h5>Zusätzliche Filter zu {}</h5>".format(
+                                    _(form_labels["ausbildung_typ_%s" % ausbildungstyp.lower()])
+                                )
+                            ),
+                            Row(
+                                *[
+                                    Column(
+                                        button_group_filter(
+                                            "ausbildung_typ_%s_%s"
+                                            % (ausbildungstyp.lower(), f.lower())
+                                        ),
+                                        css_class="form-group",
+                                        css_id=f.replace("_", "-"),
+                                    )
+                                    for f in felder.keys()
+                                    if "ausbildung_typ_medstud_abschnitt"
+                                    == "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                                ]
+                            ),
+                            *[
+                                Column(
+                                    button_group_filter(
+                                        "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                                    ),
+                                    css_class="form-group col-md-6 mb-0",
+                                    css_id=f.replace("_", "-"),
+                                )
+                                for f in felder.keys()
+                                if "ausbildung_typ_medstud_abschnitt"
+                                != "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                            ],
+                            css_id="div-ausbildung-%s" % AUSBILDUNGS_IDS[ausbildungstyp],
+                            css_class="hidden ausbildung-addon",
+                        )
+                    ]
+                )
+            else:
+                helper.layout.extend(
+                    [
+                        Div(
+                            HTML(
+                                "<hr><h5>Zusätzliche Filter zu {}</h5>".format(
+                                    _(form_labels["ausbildung_typ_%s" % ausbildungstyp.lower()])
+                                )
+                            ),
+                            Row(
+                                *[
+                                    Column(
+                                        button_group_filter(
+                                            "ausbildung_typ_%s_%s"
+                                            % (ausbildungstyp.lower(), f.lower())
+                                        ),
+                                        css_class="form-group",
+                                        css_id=f.replace("_", "-"),
+                                    )
+                                    for f in felder.keys()
+                                    if "ausbildung_typ_medstud_abschnitt"
+                                    == "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                                ]
+                            ),
+                            HTML("<p>"),
+                            HTML(
+                                _(
+                                    "In welchen der folgenden Bereiche sind Vorerfahrungen notwendig?"
+                                )
+                            ),
+                            HTML("</p>"),
+                            *[
+                                Column(
+                                    button_group_filter(
+                                        "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                                    ),
+                                    css_class="form-group col-md-6 mb-0",
+                                    css_id=f.replace("_", "-"),
+                                )
+                                for f in felder.keys()
+                                if "ausbildung_typ_medstud_abschnitt"
+                                != "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
+                            ],
+                            css_id="div-ausbildung-%s" % AUSBILDUNGS_IDS[ausbildungstyp],
+                            css_class="hidden ausbildung-addon",
+                        )
+                    ]
+                )
+
+    helper.form_tag = False
+    helper.layout.extend(
+        [
+            HTML("<hr><h5>"),
+            HTML(_(" Welche Infos über die zu vergebende Stelle sind schon bekannt?")),
+            HTML("</h5>"),
+            "availability_start",
+            Row(
+                Column(InlineRadios("braucht_bezahlung")),
+                Column(InlineRadios("unterkunft_gewuenscht")),
+            ),
+        ]
+    )
+    return helper
