+++ backend/apps/iamstudent/forms/student.py
@@ -1,7 +1,5 @@
-# from django.forms import *
 import logging
 
-from crispy_forms.bootstrap import InlineRadios
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Column, Div, Field, HTML, Layout, Row, Submit
 from django import forms
@@ -11,10 +9,9 @@ from django.utils.text import format_lazy
 from django.utils.translation import gettext_lazy as _
 
 from apps.accounts.modelss import User
+from apps.iamstudent.models import Student
+from apps.iamstudent.models.student import AUSBILDUNGS_IDS, AUSBILDUNGS_TYPEN
 from apps.iamstudent.web.buttongroup import RadioButtons
-
-from .models import EmailToSend, Student
-from .models.student import AUSBILDUNGS_IDS, AUSBILDUNGS_TYPEN
 
 form_labels = {
     "uuid": _("Writerekp"),
@@ -418,11 +415,6 @@ class StudentFormAndMail(StudentForm):
     email = forms.EmailField()
 
 
-class EmailForm(forms.Form):
-    student_id = forms.CharField(max_length=100)
-    contact_adress = forms.EmailField()
-
-
 class StudentFormEditProfile(StudentForm):
     class Meta:
         model = Student
@@ -706,150 +698,3 @@ class StudentFormView(StudentForm):
                 css_class="form-row",
             ),
         )
-
-
-class EmailToSendForm(forms.ModelForm):
-    class Meta:
-        model = EmailToSend
-        fields = ["subject", "message"]
-        labels = {"subject": _("Betreff"), "message": _("Nachrichtentext")}
-        help_texts = {}
-
-    def clean_message(self):
-        message = self.cleaned_data["message"]
-        initial_message = self.initial["message"]
-        if "".join(str(message).split()) == "".join(str(initial_message).split()):
-            raise ValidationError(_("Bitte personalisiere diesen Text"), code="invalid")
-        return message
-
-
-def get_form_helper_filter():
-    helper = FormHelper()
-    helper.form_id = "id-exampleForm"
-    helper.form_class = "blueForms"
-    helper.form_method = "get"
-
-    helper.form_action = "submit_survey"
-    helper.form_style = "inline"
-    helper.layout = Layout(
-        Div(
-            Row(
-                *[
-                    Column(
-                        "ausbildung_typ_%s" % k.lower(),
-                        css_class="ausbildung-checkbox form-group col-md-6 mb-0",
-                        css_id="ausbildung-checkbox-%s" % AUSBILDUNGS_IDS[k],
-                    )
-                    for k in AUSBILDUNGS_TYPEN.keys()
-                ]
-            ),
-            css_id="div-berufsausbildung-dropdown",
-        )
-    )
-
-    for ausbildungstyp, felder in AUSBILDUNGS_TYPEN.items():
-        if len(felder) != 0:
-            if ausbildungstyp != "MEDSTUD":
-                helper.layout.extend(
-                    [
-                        Div(
-                            HTML(
-                                "<hr><h5>Zusätzliche Filter zu {}</h5>".format(
-                                    _(form_labels["ausbildung_typ_%s" % ausbildungstyp.lower()])
-                                )
-                            ),
-                            Row(
-                                *[
-                                    Column(
-                                        button_group_filter(
-                                            "ausbildung_typ_%s_%s"
-                                            % (ausbildungstyp.lower(), f.lower())
-                                        ),
-                                        css_class="form-group",
-                                        css_id=f.replace("_", "-"),
-                                    )
-                                    for f in felder.keys()
-                                    if "ausbildung_typ_medstud_abschnitt"
-                                    == "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                                ]
-                            ),
-                            *[
-                                Column(
-                                    button_group_filter(
-                                        "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                                    ),
-                                    css_class="form-group col-md-6 mb-0",
-                                    css_id=f.replace("_", "-"),
-                                )
-                                for f in felder.keys()
-                                if "ausbildung_typ_medstud_abschnitt"
-                                != "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                            ],
-                            css_id="div-ausbildung-%s" % AUSBILDUNGS_IDS[ausbildungstyp],
-                            css_class="hidden ausbildung-addon",
-                        )
-                    ]
-                )
-            else:
-                helper.layout.extend(
-                    [
-                        Div(
-                            HTML(
-                                "<hr><h5>Zusätzliche Filter zu {}</h5>".format(
-                                    _(form_labels["ausbildung_typ_%s" % ausbildungstyp.lower()])
-                                )
-                            ),
-                            Row(
-                                *[
-                                    Column(
-                                        button_group_filter(
-                                            "ausbildung_typ_%s_%s"
-                                            % (ausbildungstyp.lower(), f.lower())
-                                        ),
-                                        css_class="form-group",
-                                        css_id=f.replace("_", "-"),
-                                    )
-                                    for f in felder.keys()
-                                    if "ausbildung_typ_medstud_abschnitt"
-                                    == "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                                ]
-                            ),
-                            HTML("<p>"),
-                            HTML(
-                                _(
-                                    "In welchen der folgenden Bereiche sind Vorerfahrungen notwendig?"
-                                )
-                            ),
-                            HTML("</p>"),
-                            *[
-                                Column(
-                                    button_group_filter(
-                                        "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                                    ),
-                                    css_class="form-group col-md-6 mb-0",
-                                    css_id=f.replace("_", "-"),
-                                )
-                                for f in felder.keys()
-                                if "ausbildung_typ_medstud_abschnitt"
-                                != "ausbildung_typ_%s_%s" % (ausbildungstyp.lower(), f.lower())
-                            ],
-                            css_id="div-ausbildung-%s" % AUSBILDUNGS_IDS[ausbildungstyp],
-                            css_class="hidden ausbildung-addon",
-                        )
-                    ]
-                )
-
-    helper.form_tag = False
-    helper.layout.extend(
-        [
-            HTML("<hr><h5>"),
-            HTML(_(" Welche Infos über die zu vergebende Stelle sind schon bekannt?")),
-            HTML("</h5>"),
-            "availability_start",
-            Row(
-                Column(InlineRadios("braucht_bezahlung")),
-                Column(InlineRadios("unterkunft_gewuenscht")),
-            ),
-        ]
-    )
-    return helper
