--- backend/apps/iamstudent/forms_filter.py
@@ -1,13 +1,0 @@
-from django import forms
-from django.utils.translation import gettext_lazy as _
-
-from .forms import form_labels
-from .models import StudentListFilterModel
-
-
-class StudentListFilterModelForm(forms.ModelForm):
-    class Meta:
-        model = StudentListFilterModel
-        labels = form_labels
-        labels["braucht_bezahlung"] = _("Vergütung möglich")
-        exclude = []
